remove casts to nondet_int
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@3697 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ test/programs/nondet_init_programs/ssh-simplified/s3_srvr_1_BUG.cil.c
@@ -576,7 +576,7 @@ int ssl3_accept(int initial_state )
                                                                             goto switch_1_break;
                                                                             switch_1_8448: 
 #line 428
-                                                                            num1 = (long )nondet_int();
+                                                                            num1 = nondet_int();
 #line 429
                                                                             if (num1 > 0L) {
 #line 430
